# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/books/create": {
      "post": {
        "operationId": "BookController_create",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Livre créé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer un nouveau livre",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/list": {
      "get": {
        "operationId": "BookController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister tous les livres de l'utilisateur",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/share/{token}": {
      "get": {
        "operationId": "BookController_accessByShareToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Token de partage",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accéder à un livre via lien de partage",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/slug/{slug}": {
      "get": {
        "operationId": "BookController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Slug du livre",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer un livre par son slug",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{id}": {
      "get": {
        "operationId": "BookController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer un livre par ID (avec chapitres, TDM, collaborateurs)",
        "tags": [
          "Books"
        ]
      },
      "patch": {
        "operationId": "BookController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un livre",
        "tags": [
          "Books"
        ]
      },
      "delete": {
        "operationId": "BookController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un livre (soft delete)",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{id}/publish": {
      "patch": {
        "operationId": "BookController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publier un livre",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{id}/archive": {
      "patch": {
        "operationId": "BookController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archiver un livre",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{id}/visibility": {
      "patch": {
        "operationId": "BookController_updateVisibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVisibilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changer la visibilité du livre (génère un lien de partage si RESTRICTED)",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{id}/chapters": {
      "post": {
        "operationId": "BookController_createChapter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du livre",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChapterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajouter un chapitre au livre",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{id}/chapters/{chapterId}": {
      "patch": {
        "operationId": "BookController_updateChapter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du livre",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chapterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChapterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour un chapitre",
        "tags": [
          "Books"
        ]
      },
      "delete": {
        "operationId": "BookController_deleteChapter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chapterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un chapitre",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{id}/chapters/2/reorder": {
      "patch": {
        "operationId": "BookController_reorderChapters",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Réordonner les chapitres",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{bookId}/chapters/{chapterId}/images": {
      "post": {
        "operationId": "BookController_createChapterImage",
        "parameters": [
          {
            "name": "bookId",
            "required": true,
            "in": "path",
            "description": "ID du livre",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chapterId",
            "required": true,
            "in": "path",
            "description": "ID du chapitre",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChapterImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image ajoutée"
          },
          "403": {
            "description": "Accès refusé"
          },
          "404": {
            "description": "Livre ou chapitre introuvable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajouter une image à un chapitre",
        "tags": [
          "Books"
        ]
      },
      "get": {
        "operationId": "BookController_findAllChapterImage",
        "parameters": [
          {
            "name": "bookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chapterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister toutes les images d'un chapitre",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{bookId}/chapters/{chapterId}/images/reorder": {
      "patch": {
        "operationId": "BookController_reorderChapterImage",
        "parameters": [
          {
            "name": "bookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chapterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Réordonner les images du chapitre",
        "tags": [
          "Books"
        ]
      }
    },
...
    "/api/v1/books/{id}/chapters/{chapterId}/sub-chapters": {
      "post": {
        "operationId": "BookController_createSubChapter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chapterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubChapterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajouter un sous-chapitre",
        "tags": [
          "Books"
        ]
      }
    },
...
  },
  "info": {
    "title": "projet-x-service",
    "description": "API du projet-x-service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateBookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Ma vie en Côte d'Ivoire"
          },
          "subtitle": {
            "type": "string",
            "example": "Une histoire de résilience"
          },
          "description": {
            "type": "string",
            "example": "Ce livre retrace..."
          },
          "language": {
            "type": "string",
            "example": "fr"
          },
          "genre": {
            "type": "string",
            "enum": [
              "AUTOBIOGRAPHY",
              "BIOGRAPHY",
              "FAMILY_ARCHIVE",
              "COMMUNITY_HISTORY",
              "MEMOIR",
              "OTHER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_PROGRESS",
              "REVIEW",
              "PUBLISHED",
              "ARCHIVED"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "RESTRICTED",
              "PUBLIC"
            ]
          },
          "targetWordCount": {
            "type": "number",
            "example": 50000
          },
          "targetPageCount": {
            "type": "number",
            "example": 200
          },
          "coverImageUrl": {
            "type": "string",
            "example": "https://cdn.example.com/cover.jpg"
          },
          "coverStyle": {
            "type": "string",
            "example": "vintage"
          },
          "isPremium": {
            "type": "boolean",
            "description": "Marquer comme premium",
            "example": false
          }
        }
      },
...
    }
  }
}